Extract some functions to `bar.py` and `energy_decomposition.py`
